[Monk] Attempt to fix clang errors

+++ engine/class_modules/apl/apl_monk.cpp
@@ -579,7 +579,7 @@ void windwalker( player_t* p )
         cd_serenity->add_action( "use_item,name=" + item.name_str + ",if=variable.serenity_burst|fight_remains<20" );
       else if ( item.name_str == "shadowgrasp_totem" )
         cd_serenity->add_action( "use_item,name=" + item.name_str + ",if=pet.xuen_the_white_tiger.active|fight_remains<20|!runeforge.invokers_delight" );
-      else if ( item.name_str.find( "gladiators_badge" ) != -1 )
+      else if ( (int)item.name_str.find( "gladiators_badge" ) != -1 )
         cd_serenity->add_action( "use_item,name=" + item.name_str + ",if=variable.serenity_burst|fight_remains<20" );
       else if ( item.name_str == "the_first_sigil" )
         cd_serenity->add_action( "use_item,name=" + item.name_str + ",if=variable.serenity_burst|fight_remains<20" );
@@ -662,7 +662,7 @@ void windwalker( player_t* p )
       else if ( item.name_str == "overcharged_anima_battery" )
         cd_sef->add_action( "use_item,name=" + item.name_str +
                             ",if=pet.xuen_the_white_tiger.active|cooldown.invoke_xuen_the_white_tiger.remains>90|fight_remains<20" );
-      else if ( item.name_str.find( "gladiators_badge" ) != -1 )
+      else if ( (int)item.name_str.find( "gladiators_badge" ) != -1 )
         cd_sef->add_action( "use_item,name=" + item.name_str +
                             ",if=covenant.necrolord&(buff.storm_earth_and_fire.remains>10|buff.bonedust_brew.up&cooldown.bonedust_brew.remains>30)|!covenant.necrolord&(cooldown.invoke_xuen_the_white_tiger.remains>55|variable.hold_xuen)|fight_remains<15" );
       else if ( item.name_str == "the_first_sigil" )
